requirements.txt
beautifulsoup4==4.11.1
python-decouple==3.6
requests==2.28.1
peewee==3.15.2
psycopg2-binary==2.9.3